PurchasesAreCompletedBy+HybridAdditions + make recordPurchase failure earlier with no product
