+++ ios/PurchasesHybridCommon/PurchasesHybridCommonIntegrationTests/StoreKitIntegrationTests.swift
@@ -30,9 +30,11 @@ class StoreKit2ObserverModeIntegrationTests: BaseIntegrationTests {
     override class var storeKitVersion: StoreKitVersion { return .storeKit2 }
 
     func testRecordPurchase() async throws {
-        let product = try await Product.products(for: [Self.productIdentifier]).first
-        expect(product).toNot(beNil(), description: "The product that we're going to purchase should not be nil.")
-        let purchaseResult = try await product?.purchase()
+        guard let product = try await Product.products(for: [Self.productIdentifier]).first else {
+            fail("The product to purchase should not be nil.")
+            return
+        }
+        let purchaseResult = try await product.purchase()
         expect(purchaseResult).toNot(beNil(), description: "The purchase result should not be nil.")
 
         let (dict, error) = try await withCheckedThrowingContinuation { continuation in
@@ -41,7 +43,10 @@ class StoreKit2ObserverModeIntegrationTests: BaseIntegrationTests {
             }
 
         }
-        expect(error).to(beNil(), description: "The error returned by recordPurchase() should be nil.")
+        expect(error).to(
+            beNil(),
+            description: "The error returned by recordPurchase() should be nil, got \(error.debugDescription)."
+        )
         var unwrappedDict = try XCTUnwrap(dict)
         removeDates(&unwrappedDict)
         await self.assertSnapshot(unwrappedDict)
